Level 3 . 제네릭,람다,스트림

+++ src/Level2/Arithmetic/ModOperator.java
@@ -1,7 +1,7 @@
 package Level2.Arithmetic;
 
 public class ModOperator implements InterOperator{
-    public int operate(int a, int b) {
+    public double operate(int a, int b) {
         return a % b;
     }
 }
